% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/short_term.R
\name{add_dates}
\alias{add_dates}
\title{helper function to add dates to the projected time points}
\usage{
add_dates(df, date, look_ahead)
}
\arguments{
\item{df}{model output df}

\item{date}{sys.date()}

\item{look_ahead}{number of days to forecast: default is 14}
}
\description{
helper function to add dates to the projected time points
}
