% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtest.R
\name{backtest_short}
\alias{backtest_short}
\title{function to backtest model parameters.}
\usage{
backtest_short(ticker, n_tests, ncores, vendor = "quandl", ...)
}
\arguments{
\item{ticker}{which ticker or tickers should we evaluate?}

\item{n_tests}{how many tests per ticker per month of data should we use?}

\item{ncores}{number of cores to use for parallel computing.}

\item{...}{additional parameters passed to renarin_short_historic}
}
\description{
compare model forecasts to the efficient market hypothesis
(i.e. assuming share price is the same the next day as it was the previous day)
}
