% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_reference.R
\name{clean_fundamentals}
\alias{clean_fundamentals}
\title{internal function to help process fundamentals data}
\usage{
clean_fundamentals(df, year)
}
\arguments{
\item{df}{data frame}

\item{year}{year or range of years to create the reference database from.
good idea to choose range with similar macroeconomic conditions to current economy.}
}
\description{
This is filtering out non-quarterly reports and adjusted reports, limiting to a certain year or range of years, and
}
