% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_reference.R
\name{create_reference}
\alias{create_reference}
\title{wrapper around tidyquant functions to create a reference dataset for our forecasting system}
\usage{
create_reference(
  vendor = "quandl",
  ncores = 4,
  data_repo = "G:/My Drive/pkg/data/sharadar/",
  year = 2018
)
}
\arguments{
\item{vendor}{which data warehouse - currently only works with the Sharadar tables from quandl}

\item{ncores}{number of cores to utilize for data processing and ingestion}

\item{data_repo}{filepath where function can find the Sharadar/SF1 table,}

\item{year}{year or range of years to create the reference database from.
good idea to choose range with similar macroeconomic conditions to current economy.}
}
\value{
tibble containing time series data for our analysis.
}
\description{
was going to set this up entirely via API but turns out the easiest way to do this by far is to download the full
tables via batch download and start from there
}
