% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_timevec}
\alias{generate_timevec}
\title{function to generate vector of start times for a given ticker}
\usage{
generate_timevec(ticker, df, num_times)
}
\arguments{
\item{ticker}{security identifier}

\item{df}{dataframe containing historical data about security price}

\item{num_times}{number of timeframes we want per calendar month (defaults to one)}
}
\description{
helper function for historical performance of forecasting system.
}
