% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_reference.R
\name{get_api_key}
\alias{get_api_key}
\title{internal function (me only) to return API key for different data vendors}
\usage{
get_api_key(vendor = "quandl")
}
\arguments{
\item{vendor}{which data warehouse - currently only works with the Sharadar tables from quandl}
}
\description{
internal function (me only) to return API key for different data vendors
}
