% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_sp500}
\alias{get_sp500}
\title{function to query wikipedia and pull metadata regarding tickers on the S and P 500}
\usage{
get_sp500(keep_metadata = FALSE)
}
\arguments{
\item{keep_metadata}{logical. should we keep metadata?}
}
\description{
what the title says
}
