% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtest.R
\name{get_times}
\alias{get_times}
\title{generate vector of timevecs for a given ticker}
\usage{
get_times(ticker, n_tests, vendor)
}
\arguments{
\item{ticker}{which ticker or tickers should we evaluate?}

\item{n_tests}{how many tests per ticker per month of data should we use?}
}
\description{
helper function for backtest_short - this function only works with the Sharadar
data feed.
}
