% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/short_term_historic.R
\name{process_fit_historic}
\alias{process_fit_historic}
\title{helper function to process and join forecast output from renarin_short}
\usage{
process_fit_historic(fit, fcast, df, PI)
}
\arguments{
\item{fit}{model fit object (from nnetar)}

\item{fcast}{model forecast object}

\item{df}{model df}
}
\description{
helper function to process and join forecast output from renarin_short
}
