% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/short_term.R
\name{summarize_forecast}
\alias{summarize_forecast}
\title{function to summarize the output of short-term model (for back-testing)}
\usage{
summarize_forecast(df)
}
\arguments{
\item{df}{dataframe containing projections for a given time period}
}
\description{
function to summarize the output of short-term model (for back-testing)
}
