% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/short_term_historic.R
\name{summarize_forecast_historic}
\alias{summarize_forecast_historic}
\title{function to summarize the output of short-term model (for back-testing)}
\usage{
summarize_forecast_historic(df, PI)
}
\arguments{
\item{df}{dataframe containing projections for a given time period}
}
\description{
function to summarize the output of short-term model (for back-testing)
}
