% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/short_term.R
\name{undiff}
\alias{undiff}
\title{helper function to undiff forecasts from renarin_short -}
\usage{
undiff(x, y, z)
}
\arguments{
\item{x}{vector of differenced values to undiff}

\item{y}{vector of reference values}

\item{z}{vector of indices.}
}
\description{
tricky because the differences were lagged by one but the forecast lags by 10.
}
