% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/short_term.R
\name{renarin_short}
\alias{renarin_short}
\title{function to generate forecast for a given ticker}
\usage{
renarin_short(ticker, vendor = "quandl", look_back = 90, look_ahead = 7)
}
\arguments{
\item{ticker}{security name, e.g. AAPL}

\item{vendor}{provider for data}

\item{look_back}{number of days to look back for predictions}

\item{look_ahead}{number of days to forecast}
}
\description{
use nnet forecasting to evaluate future stock price.
}
