% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/short_term_historic.R
\name{renarin_short_historic}
\alias{renarin_short_historic}
\title{function to generate forecast for a given ticker (using historical data).}
\usage{
renarin_short_historic(
  date = NULL,
  ticker,
  look_back = 200,
  look_ahead = 14,
  zoom_in = FALSE,
  lag = 20,
  decay = 0.2,
  vendor = "quandl",
  PI = FALSE
)
}
\arguments{
\item{date}{if most recent == FALSE, what date should we project from?}

\item{ticker}{security name, e.g. AAPL}

\item{look_back}{number of days to look back for predictions}

\item{look_ahead}{number of days to forecast}

\item{zoom_in}{should function return full model and forecast results? useful for highlighting speicific issues}

\item{lag}{number of days to use to fit the next day in the model}

\item{decay}{NNET parametar}

\item{df}{if most_recent = FALSE, provide historical pricing dataframe}
}
\description{
use nnet forecasting to evaluate future stock price.
}
