% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/short_term.R
\name{renarin_short_historic}
\alias{renarin_short_historic}
\title{function to generate forecast for a given ticker (using historical data).}
\usage{
renarin_short_historic(
  ticker,
  date = NULL,
  df = NULL,
  look_back = 30,
  look_ahead = 7,
  zoom_in = FALSE
)
}
\arguments{
\item{ticker}{security name, e.g. AAPL}

\item{date}{if most recent == FALSE, what date should we project from?}

\item{df}{if most_recent = FALSE, provide historical pricing dataframe}

\item{look_back}{number of days to look back for predictions}

\item{look_ahead}{number of days to forecast}

\item{zoom_in}{should function return full model and forecast results? useful for highlighting speicific issues}
}
\description{
use nnet forecasting to evaluate future stock price.
}
