% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/short_term.R
\name{summarize_forecast}
\alias{summarize_forecast}
\title{helper function to summarize forecast}
\usage{
summarize_forecast(df)
}
\arguments{
\item{df}{model output df}
}
\description{
helper function to summarize forecast
}
